Bugfix for single BBL RPs
Single BBL RPs did not correctly switch back to normal execution after a return
instruction. The reason is that BBL_SetTargetVersion() was called twice on the
same BBL (the only one in the RP).

We now use INS_InsertVersionCase() to switch on the first instruction of a RP,
which immediately switches versions, and the 1st BBL is instrumented under the
checkpointing version. On the other hand, returns are instrumented using
BBL_SetTargetVersion() to cause the *next* BBL to be executed in normal
execution mode.

Also, added some more descriptive messages for installing checkpoint entries and
exits.