Fix path of encoded bins inserted in db
The path inserted in the db is also obtained after calling readlink -f.
Searching the key of an image already occurs this way. i

This solves the error where the key of an image is not found because the insert
and the select do not use the same absolute path for the same file.